Creates CategoryListView and associated items with it
we introduced some mockups and now that has changed some ways we design stuffs. This is the first iteration of said changes. We are creating a categorylist component with toolbar and lot more functionalities. The warning will address each step of the toolbar during the process. There are tooltip components to guide the user as well.

The color scheme used is still default to material ui. I have not added any of our native color choices yet. All components are only renedered in the storybook component. You need to storybook to view the components as the landing page for them have not yet been created yet.