Adds new layout and styles (#31)
- adds app bar to display navigation instead of nav bar. we are doing this to make it more uniform on a mobile view.
- words changed in idea component
- cleanup on layout.js component to remove navbar and add appbar
- applying inner padding for roi calculator but this needs to be redesigned
- cleanup on subscribe component as users can save first name and last name

at the moment there is no requirement to have a user matching profile to a subscribe profile. however, there must be a link associated with it. we might need to think how we are going to apply roles and permissions if the user is not a valid user. at the moment without a validity it is not possible to view categories and item lists.