apply ability to view dashboard easily
Introducing the ability to view the dashboard page without being logged in. This allows the users ability to view all active properties that are for rent without even logging in. This feature allows for page level layout from nextjs and controls the layout of each page. By said functionality we are able to allow unsigned users to also view properties.