Revamp profile details page
- remove constant error handling for adding an inventory. we had
  duplicate values so removed the unused ones.
- update appearance settings ux to use light vs dark theme. this needs
  more work because it doesn't switch the theme at all. we have not
created a dark mode / theme for this either.
- update notification handler for profile page to support tanstack
  query. we used query client to invalidate data when adding a new item
or change is detected. this allows the lib to refetch required data so
its always fresh on the client side.
- add mutations fn and getQuery fn for profile details page. i think we
  should refactor this later on because there are already four or more
files and we are going to be adding more. i tried to group them together
but for some reason i'm not able to group them together. need more
investigation on that.
- tech debt and css cleanup.
closes #122