Revamp structure to support faster rendering (#135)
- moving away from nextjs as we want faster client side routing

with the existing use of pages and app components, nextjs is pretty slow and we also
do not have the need to support ssr or enable server side components. due to this requirements,
it is not really required to move towards nextjs and because of this we just snuck in vite config
and used it as designed. Existing components are just the same. Changed a few names and cleaned up
some of the existing bugs.

- cleanup package contents and removed all unwanted config files

there are some stuff still under containers and i would like to move them toward components since with
the use of tanstack query we really do not need any redux actions, so henceforth do not need any of the
container component logic to prevail. why makes this difficult when we can make it easy.

- add support for dark mode and add ability to update profile details

this was one of the biggest changes. with the advent of dark mode, the users are now able to view things
in a much more pleasant fashion. used different theme for dark mode, we can adjust as we go if needed.