130 - Add support for bulk inventory upload
Add support for bulk upload of inventory items
- remove package lock since we are using yarn
- remove on success mutation from component files
- support bulk upload of inventory
- support for key, props in autocomplete tech debt error fix for react
- remove unused react state and add boolean for inventory field during edit
- add confirmation dialog box for creating / updating inventory and / or category and maintenance plans
- handle updated and created time differently for profile page
- add rtk for adding bulk inventories

Add relationship for category, maintenance and inventory items

- add support for category and item
- add support for maintenance plan and item
- add new column for item description, removes useless tooltip
If inventory item is assigned to a specific plan, then only we allow for tooltip

- add ability to assign single or multiple items to a selected maintenance
plan.
- update counts for total items that fall under a certain plan
- add sql support to update plan and associated maintenance items
- add ability to update and assign categories to inventories
- items that are under a select plan can be view when the user clicks on the Total items tab.
this gives the user the ability to view all inventory items associated against that selected maintenance plan.

- add chart for better visualization as well.
- add color to maintenance plan
- add mui-color-input package lib for selecting colors
- add comments for InventoryTable
we are using this file heavily to display tables so added
some top level comments so that its easier to read and understand
the usage and purpose of this table.

- added variants to inventory table
added some variants to this file, so that it behaves differently
in different scenarios

- added ability to attach custom color for selected maintenance plans
added ability to attach colors to select maintenance plan. added trigger to associate
each inventory item against a color and populate the color based on the maintenance plan

- minor sql changes
added some minor sql changes that were tech debt and needed to be cleared off

- add due date for maintenance items according to the selection of the user.
this due date currently is used to display when the next inspection is due. Using pg_cron to determine
the current timestamp and select the best / nearest option for the due date. every day at night, the cron is
scheduled to run. instead of occupying any database operations, we decided to use this pg_cron tool so that we
could easily update / edit the db records. Similar strategy would be applied when the records get too much and need
some cleanup. There might be some preventive actions that are required to get this setup working

Closes #130