Add support for assigning items to Categories
- added new route to support category list
there are a lot more ways to add category, but to give users a streamlined
experience, we are engaging into this nature of workflow. this is similar to
how the maintenance plans can be created / edited.
- added color support with each category however the usage has not yet been
discovered yet. we will add the use of color for each category once we get there