Add ability to manipulate inventory items
- remove package.lock and yarn.lock
a lot of times, we were using one or the other. due to that it started
to become a lot more hectic to manage that. i think the best option here
is to let the devs choose what they want to use, and refrain from
pushing such changes.

- added default formatter as prettier
this seems to be the standard anyways. so until we find a better
alternative, we should keep it at that. slowly we should add more eslint
capabilities as it does strengthen the dev team and encourages devs to
write proper code.

- added loadash package to support debounce fn
i tried to use our own custom debounce fn but that was not working as i
thought with rtk/tanstack. due to that i thought it would be best if we
could just use what is provided. in the near future, if possible we can
move to our custom implementation but that doesn't seem like a need atm.

- added the ability to flex wrap the categories component. this gives us
  cleaner look when going through the categories. updated code to
support deletion of categories and added confirmation box for this.

- added more meaningful constants and helper messages for charts and
  some other components. this gives the user the ability to know what is
going on throughout the page.

- added support for pie chart for product detail widget view. i have
  just rough handled the code atm, didn't think we wanted to optimize at
the start since there are more unknowns and trying to prematurely
optimize is waste of dev resource. when the need comes, we will update
the code to be more reasonable and flush.

- added new components to track category frequency and its own
  associated limit. this gave me the most difficult time to get across
since each item can have their own associated limits. i was
over-engineering it a lot and after a couple of tries i was able to get
it into the state that i would like it to be at. the checkbox at the end
denotes what is selected and if its not selected then we assume the
input is selected. since the backend does not store the selection menu
name only the value there was a lot of map and filter that we needed to
use. sometimes i think maybe category should just have stored the
selected threshold name and value instead of just the value. it would
have made the ui more streamlined that is for sure.

- created a base structure to display items that fall under the selected
  threshold of each category. since there is backend work to be done
here, i didn't think it was ok to mess with the logic since this pr is
also very huge already. it would be difficult to keep lean if we do not
actuate proper judgements to keep the content pristine. will work on
this in the upcomming ticket.

- added routes for user interaction for the home icon and for individual
  primary bar icons / menu items. previously we didn't have anything of
that nature, its now good to move around faster.

- added tabbed component to support forecase of items and categories.

- cleanup on inventories and added plainView. this flag is used to not
  display checkboxes but if unselected, checkboxes are displayed where
the user can select all of the inventories and quickly assign them to
categories or maintenance plans.

- added inventory overview that gives clients a brief idea of what they
  are / can see in the landing page. this gives an overview of all the
categories, plan, and also provides some graphical representation of
items.

- cleanup on existing tanstack queries for easy to read and proper
  access.

 - added thresholdlimit to the category sql page

- tech debt and cleanup for css and removes un-used files. also
  attempted to use proper variable names. i know its not a well done
job, but if there are pr comments to address that, i will certainly
update code for that.

- modify bookmarked view to view items recently edited, capability for bookmarked will be used later on
- remove grid in categories component to support full width for card components
- add cssbaseline to support dark mode all around
- add support for numeric text field in select quantity during forecast

closes #139