Add low stock page and refactor table component
- remove unused files for header with button component
there were a lot of extra props that I think was not required and hence
added the cleanup for those files. we were able to reduce the component
size, but i think we could further reduce it but it should suffice for
now.

- populate table for low stocks items
refactored quite a bit of code since it was not properly formatted. this
reduced a lot of boiler plate from the file as well. added button group
to handle the grouping of assign and delete for easier access.

- adds suspense and lazy loading
adds suspense and lazy loading to the router files. this helps with
chunk loading and due to this decreases the load time. i was running
into an issue where the page would not properly render but that was due
to not having a suspense. Added a suspence item and also added div to
support the layout.

- add variation to table component
removed a couple of un-used files and refactored the code quite a bit

- adds support to build out the ability to search on list of inventory
  items
added a small search that allows searching for a select item. During
this I noticed that the component TableComponent is getting difficult
to manage. So we might need to streamline this. We have requirements
that need to support grid mode, list mode and a lot of other stuffs. The
inventory list table is also used to display items that belong to a
certain category and we need to be able to move items from one category
to another. this is also a must for maintenance plans and before we
start digging in too deep, its best that we fix pending issues before
then.

- add support to generate qr code for inventory items
the qr code currently only reads in the inventory item. a potential
benefit would also allow you to integrate this with categories or
maintenance plans, so you can see the count of items that fall under the
category / plan.

- adds grid layout for inventory items
support for the ability to view items in grid vs list layout. users can
save the preference in thier profile settings. default mode is list
view. only is used in the inventory list component. other places do not
allow for grid layout. i tried to add the default image, but for some
reason the svg is not being applied properly. need to investigate why,
the bg.svg is applied but the canvas fails to be applied.

- adds dist to gitignore
noticed that if we would build, then the dist folder was not added to
git ignore so i added that as well.

- repositioned the autocomplete out of inventory table
decided to move autocomplete out of the table component since it is not
strictly in relation to the table component. i also want to refrain from
having ton of inventories and managing state in the same component. it
will quickly get bloated. this supports search for grid layout too which
is a nice to have.

- tech debt and cleanup

closes #144