Add warning banner for found inventories
adds warning banner if selected inventories is found assigned to any of the
existing categories or maintenance plans. instead of reassigning it, we
provide users the ability to add same item in many different categories
and plans so that users can have refined control of the selected item.
if the user wants to remove it from a certain category, users must
utlize the selected category tab or maintenance tab to do such.