149 move items (#151)
* Add warning banner for found inventories

adds warning banner if selected inventories is found assigned to any of the
existing categories or maintenance plans. instead of reassigning it, we
provide users the ability to add same item in many different categories
and plans so that users can have refined control of the selected item.
if the user wants to remove it from a certain category, users must
utlize the selected category tab or maintenance tab to do such.

* Remove inventory item from selection

- adds ability to remove inventory item from selection menu. if a
  inventory is assigned to a category, users can remove the selection
manually by clicking the restore icon that allows users to restore the
selection. similar concept with maintenance plan.

- cleanup on table component so that its easier to build this feature
  out. extracted the rowsFormatter fn so that its easier to work with
the row selection based on where the user is in the page.

closes #149

* move svg files in public

* refactor components