Add optimistic updates for portions of app
- adds optimistic updates for certain a/o of the app. this supports
  easy rendering of change in the ui.

- add edit icons to update price and quantity so that the user does not
  have to click multiple times to edit a single item. this supports
quick edit which will help us build the visualizer tool for all
historical data for a selected item.

- found out the issue with isLoading flag might not re-render items
  correctly because if the state is not removed, then the fetch action
is only reproduced which in turns does not allow for the data to be
re-rendered, since we have an useEffect for that action.

- noticed that we were looking at the wrong side of the tracks because
  the error was with how the useEffect was not triggering with the
isLoading flag because technically the item was just being re-fetched.
we need the useEffect to support the load of searchable items therefore
it was preventing us from finding the real bug.